PID Controll

+++ lib/Common/PID_v1.h
@@ -18,34 +18,34 @@ class PID
 
   //commonly used functions **************************************************************************
     PID();
-    PID(float, float, float, int, int);   // * constructor.  links the PID to the Input, Output, and 
+    PID(double, double, double, int, int);   // * constructor.  links the PID to the Input, Output, and 
                                           //   Setpoint.  Initial tuning parameters are also set here.
                                           //   (overload for specifying proportional mode)
 
-    PID(float, float, float);           // * constructor.  links the PID to the Input, Output, and 
+    PID(double, double, double);           // * constructor.  links the PID to the Input, Output, and 
                                              //   Setpoint.  Initial tuning parameters are also set here
-    PID(float, float, float, int); 
+    PID(double, double, double, int); 
 	
     // void SetMode(int Mode);               // * sets PID to either Manual (0) or Auto (non-0)
 
-    float Compute(float currentValue, 
-                  float setPoint);        // * performs the PID calculation.  it should be
+    double Compute(double currentValue, 
+                  double setPoint);        // * performs the PID calculation.  it should be
                                           //   called every time loop() cycles. ON/OFF and
                                           //   calculation frequency can be set using SetMode
                                           //   SetSampleTime respectively
 
-    void SetOutputLimits(float, float);   // * clamps the output to a specific range. 0-255 by default, but
+    void SetOutputLimits(double, double);   // * clamps the output to a specific range. 0-255 by default, but
 										                      //   it's likely the user will want to change this depending on
 										                      //   the application
 	
 
 
   //available but not commonly used functions ********************************************************
-    void SetTunings(float, float,       // * While most users will set the tunings once in the 
-                    float);         	  //   constructor, this function gives the user the option
+    void SetTunings(double, double,       // * While most users will set the tunings once in the 
+                    double);         	  //   constructor, this function gives the user the option
                                         //   of changing tunings during runtime for Adaptive control
-    void SetTunings(float, float,       // * overload for specifying proportional mode
-                    float, int);         	  
+    void SetTunings(double, double,       // * overload for specifying proportional mode
+                    double, int);         	  
 
 	  void SetControllerDirection(int);	  // * Sets the Direction, or "Action" of the controller. DIRECT
 										                    //   means the output will increase when error is positive. REVERSE
@@ -57,9 +57,9 @@ class PID
 										  
 										  
   //Display functions ****************************************************************
-	float GetKp();						  // These functions query the pid for interal values.
-	float GetKi();						  //  they were created mainly for the pid front-end,
-	float GetKd();						  // where it's important to know what is actually 
+	double GetKp();						  // These functions query the pid for interal values.
+	double GetKi();						  //  they were created mainly for the pid front-end,
+	double GetKd();						  // where it's important to know what is actually 
 	// int GetMode();						  //  inside the PID.
 	int GetDirection();					//
   unsigned long windowStartTime;
@@ -67,27 +67,27 @@ class PID
   private:
 	void Initialize();
 	
-	float dispKp;				        // * we'll hold on to the tuning parameters in user-entered 
-	float dispKi;				        //   format for display purposes
-	float dispKd;				        //
+	double dispKp;				        // * we'll hold on to the tuning parameters in user-entered 
+	double dispKi;				        //   format for display purposes
+	double dispKd;				        //
     
-	float kp;                   // * (P)roportional Tuning Parameter
-  float ki;                   // * (I)ntegral Tuning Parameter
-  float kd;                   // * (D)erivative Tuning Parameter
+	double kp;                   // * (P)roportional Tuning Parameter
+  double ki;                   // * (I)ntegral Tuning Parameter
+  double kd;                   // * (D)erivative Tuning Parameter
 
 	int controllerDirection;
 	int pOn;
 
-  // float *myInput;              // * Pointers to the Input, Output, and Setpoint variables
-  // float *myOutput;             //   This creates a hard link between the variables and the 
-  // float *mySetpoint;           //   PID, freeing the user from having to constantly tell us
+  // double *myInput;              // * Pointers to the Input, Output, and Setpoint variables
+  // double *myOutput;             //   This creates a hard link between the variables and the 
+  // double *mySetpoint;           //   PID, freeing the user from having to constantly tell us
                                //   what these values are.  with pointers we'll just know.
 			  
 	unsigned long lastTime;
-	float outputSum, lastInput;
+	double outputSum, lastInput;
 
 	unsigned long SampleTime;
-	float outMin, outMax;
+	double outMin, outMax;
 	bool pOnE;
 };
 #endif
